Seção 9 - DTOs e Data Annotation

+++ Back/src/ProEventos.API/Controllers/EventosController.cs
@@ -1,14 +1,10 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Logging;
 using ProEventos.Application.Contratos;
+using ProEventos.Application.Dtos;
 using ProEventos.Domain;
-using ProEventos.Persistence;
-using ProEventos.Persistence.Contextos;
 
 namespace ProEventos.API.Controllers
 {
@@ -31,7 +27,8 @@ namespace ProEventos.API.Controllers
            try
            {
                 var eventos = await _eventoService.GetAllEventosAsync(true);
-                if (eventos == null) return NotFound("Nenhum evento encontrado!");
+                if (eventos == null) return NoContent();
+
 
                 return Ok(eventos);
            }
@@ -49,7 +46,7 @@ namespace ProEventos.API.Controllers
             try
            {
                 var evento = await _eventoService.GetEventoByIdAsync(id, true);
-                if (evento == null) return NotFound("Evento por Id não encontrado!");
+                if (evento == null) return NoContent();
 
                 return Ok(evento);
            }
@@ -67,7 +64,7 @@ namespace ProEventos.API.Controllers
             try
            {
                 var evento = await _eventoService.GetAllEventosByTemaAsync(tema, true);
-                if (evento == null) return NotFound("Eventos por Tema não encontrados!");
+                if (evento == null) return NoContent();
 
                 return Ok(evento);
            }
@@ -80,12 +77,12 @@ namespace ProEventos.API.Controllers
         }
 
         [HttpPost]
-        public async Task<IActionResult> Post(Evento model)
+        public async Task<IActionResult> Post(EventoDto model)
         {
            try
            {
                 var evento = await _eventoService.AddEventos(model);
-                if (evento == null) return BadRequest("Erro ao tentar adicionar eventos!");
+                if (evento == null) return NoContent();
 
                 return Ok(evento);
            }
@@ -98,12 +95,12 @@ namespace ProEventos.API.Controllers
         }
 
         [HttpPut("{id}")]
-        public async Task<IActionResult> Put(int id, Evento model)
+        public async Task<IActionResult> Put(int id, EventoDto model)
         {
             try
            {
                 var evento = await _eventoService.UpdateEventos(id, model);
-                if (evento == null) return BadRequest("Erro ao tentar atualizar eventos!");
+                if (evento == null) return NoContent();
 
                 return Ok(evento);
            }
@@ -120,9 +117,12 @@ namespace ProEventos.API.Controllers
         {
            try
            {
+               var evento = await _eventoService.GetEventoByIdAsync(id, true);
+               if (evento == null) return NoContent();
 
                return await _eventoService.DeleteEventos(id) ?
-                  Ok("Deletado") : BadRequest("Evento não deletado");              
+                  Ok("Deletado") : 
+                  throw new Exception("Ocorreu um problema não específico ao tentar deletar Evento");              
            }
            catch (Exception ex)
            {
